+++ et-common-libs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added CHANGELOG.md
+- Added a MM2CM message to dump the context of a thread
 ### Changed
 - (Conan) Use conan-common get_version_from_cmake_project
 ### Deprecated